requirements.txt
django
djangorestframework
gunicorn    # WSGI(WAS) server for production